Reintroduce audio
I decided that if I was going to go to the trouble of showing the Wonder Card/News on the distribution ROM's screen, I may as well include the audio. After all, in theory the character map used by the game allows for audio to be played directly from text.